Create Stefan-Boltzmann's Law
This is the GTF model that shows the final release from the Earth by the end of the Flood year using the Stefan-Boltzmann's Law, constant, and equation. Over 99% of the heat is still released by the end of the Flood year. Each zone is indivisually calculated using Stefan-Boltzmann's Law and equation in a real-world simulation of the Geothermal Feedback model, which is, currently, the most powerful model ever proposed for resollution of the so-called "heat problem" to the Global Flood.